Mostly fixed the snapping bug
It was likely due to the character being already in the wall when rotated and not getting pushed out before we snap them down to attach them to the wall. To fix it I adjusted the positioning of the sprites & collision shape so the sides and bottom are equal distance to the center of rotation.

Now it only reproduces very rarely, but to fix that I think I just need to adjust the snapping to not snap as far, and just not snap at all if it doesn't detect it colliding with anything.