First attempt to reduce jitter, needs work
This first attempt simply averages the last few angles. It reduces but will never eliminate the jitter.

I have an idea for a proper approach, which is to determine which sensor is the outlier, and simply ignore that sensor from the average. Perhaps this could be determined by if that sensor is changing the angle in the opposite direction to the other two. Or, rather than taking the average maybe take the sensor with the shortest distance to the ground.